+++ modules/tinymce/CHANGELOG.md
@@ -44,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Flash of unstyled content while loading the editor because the content css was loaded after the editor content was rendered #TINY-7249
 - Only table content would be deleted when partially selecting a table and content outside the table #TINY-6044
 - Unbinding an event handler did not take effect immediately while the event was firing #TINY-7436
+- The table cell selection handling was incorrect in some cases when dealing with nested tables #TINY-6298
 - Binding an event handler incorrectly took effect immediately while the event was firing #TINY-7436
 - Partially transparent RGBA values provided in the `color_map` setting were given the wrong hex value #TINY-7163
 - HTML comments with mismatched quotes were parsed incorrectly under certain circumstances #TINY-7589